% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{calculate_fitted_values}
\alias{calculate_fitted_values}
\title{Calculate Fitted Values for Hurdle Models}
\usage{
calculate_fitted_values(
  Y,
  X,
  Z,
  coefc,
  coefz,
  offsetx,
  offsetz,
  dist,
  zero.dist,
  theta,
  linkinv
)
}
\arguments{
\item{Y}{Response variable}

\item{X}{Count model matrix}

\item{Z}{Zero model matrix}

\item{coefc}{Count model coefficients}

\item{coefz}{Zero model coefficients}

\item{offsetx}{Count model offset}

\item{offsetz}{Zero model offset}

\item{dist}{Count distribution}

\item{zero.dist}{Zero distribution}

\item{theta}{Dispersion parameters}

\item{linkinv}{Link inverse function}
}
\value{
List with fitted values and residuals
}
\description{
Calculate Fitted Values for Hurdle Models
}
\keyword{internal}
