% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{estimate_joint_components}
\alias{estimate_joint_components}
\title{Estimate Joint Components for Hurdle Models}
\usage{
estimate_joint_components(
  Y,
  X,
  Z,
  offsetx,
  offsetz,
  weights,
  dist,
  zero.dist,
  linkstr,
  start,
  method,
  hessian,
  control,
  kx,
  kz
)
}
\arguments{
\item{Y}{Response variable}

\item{X}{Count model matrix}

\item{Z}{Zero model matrix}

\item{offsetx}{Count model offset}

\item{offsetz}{Zero model offset}

\item{weights}{Observation weights}

\item{dist}{Count distribution}

\item{zero.dist}{Zero distribution}

\item{linkstr}{Link function}

\item{start}{Starting values}

\item{method}{Optimization method}

\item{hessian}{Logical for computing hessian}

\item{control}{Control parameters}

\item{kx}{Number of count model parameters}

\item{kz}{Number of zero model parameters}
}
\value{
List of estimation results
}
\description{
Estimate Joint Components for Hurdle Models
}
\keyword{internal}
