% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{extractAIC.fasthurdle}
\alias{extractAIC.fasthurdle}
\title{Extract AIC from a Hurdle Model}
\usage{
\method{extractAIC}{fasthurdle}(fit, scale = NULL, k = 2, ...)
}
\arguments{
\item{fit}{A fitted model object of class "fasthurdle".}

\item{scale}{Scale parameter for the AIC. Not used.}

\item{k}{Penalty per parameter to be used in AIC.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A numeric vector of length 2, with the degrees of freedom and the AIC.
}
\description{
Extract AIC from a fitted hurdle model.
}
