% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{fitted.fasthurdle}
\alias{fitted.fasthurdle}
\title{Extract Fitted Values from a Hurdle Model}
\usage{
\method{fitted}{fasthurdle}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "fasthurdle".}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A vector of fitted values.
}
\description{
Extract fitted values from a fitted hurdle model.
}
