% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{generate_starting_values}
\alias{generate_starting_values}
\title{Generate Starting Values for Hurdle Models}
\usage{
generate_starting_values(
  start,
  Y,
  X,
  Z,
  offsetx,
  offsetz,
  weights,
  dist,
  zero.dist,
  linkstr,
  kx,
  kz,
  trace
)
}
\arguments{
\item{start}{User-provided starting values}

\item{Y}{Response variable}

\item{X}{Count model matrix}

\item{Z}{Zero model matrix}

\item{offsetx}{Count model offset}

\item{offsetz}{Zero model offset}

\item{weights}{Observation weights}

\item{dist}{Count distribution}

\item{zero.dist}{Zero distribution}

\item{linkstr}{Link function}

\item{kx}{Number of count model parameters}

\item{kz}{Number of zero model parameters}

\item{trace}{Logical for printing trace information}
}
\value{
List of starting values
}
\description{
Generate Starting Values for Hurdle Models
}
\keyword{internal}
