% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value.}

\item{y}{Second value.}
}
\value{
x if x is not NULL, otherwise y.
}
\description{
Return the first non-NULL value.
}
\keyword{internal}
