% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{hurdletest}
\alias{hurdletest}
\title{Test for Equality of Count and Zero Model Coefficients}
\usage{
hurdletest(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "fasthurdle".}

\item{...}{Additional arguments passed to car::linearHypothesis.}
}
\value{
An object of class "anova" containing the results of the test.
}
\description{
Test for equality of count and zero model coefficients in a hurdle model.
}
