% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{logLik.fasthurdle}
\alias{logLik.fasthurdle}
\title{Extract Log-Likelihood from a Hurdle Model}
\usage{
\method{logLik}{fasthurdle}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "fasthurdle".}

\item{...}{Additional arguments (currently ignored).}
}
\value{
An object of class "logLik".
}
\description{
Extract the log-likelihood from a fitted hurdle model.
}
