% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{model.matrix.fasthurdle}
\alias{model.matrix.fasthurdle}
\title{Extract Model Matrix from a Hurdle Model}
\usage{
\method{model.matrix}{fasthurdle}(object, model = c("count", "zero"), ...)
}
\arguments{
\item{object}{A fitted model object of class "fasthurdle".}

\item{model}{Character string specifying which model matrix to extract.
"count" extracts count model matrix, and "zero" extracts zero hurdle model matrix.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A model matrix.
}
\description{
Extract the model matrix from a fitted hurdle model.
}
