% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{model_offset_2}
\alias{model_offset_2}
\title{Extract Offset from a Model Frame}
\usage{
model_offset_2(x, terms = NULL, offset = TRUE)
}
\arguments{
\item{x}{A model frame.}

\item{terms}{Terms object. If NULL, the terms attribute of x is used.}

\item{offset}{Logical. If TRUE, the "(offset)" column is included.}
}
\value{
A numeric vector of offsets.
}
\description{
Extract offset from a model frame, optionally using different terms.
}
\keyword{internal}
