% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{predprob.fasthurdle}
\alias{predprob.fasthurdle}
\title{Predict Probabilities from a Hurdle Model}
\usage{
predprob.fasthurdle(obj, ...)
}
\arguments{
\item{obj}{A fitted model object of class "fasthurdle".}

\item{...}{Additional arguments passed to predict.fasthurdle.}
}
\value{
A matrix of predicted probabilities.
}
\description{
Predict probabilities from a fitted hurdle model.
}
