% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{print.fasthurdle}
\alias{print.fasthurdle}
\title{Print Method for Hurdle Models}
\usage{
\method{print}{fasthurdle}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A fitted model object of class "fasthurdle".}

\item{digits}{Number of significant digits to use for printing.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
The fitted model object (invisibly).
}
\description{
Print a summary of a fitted hurdle model.
}
