% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{print.summary.fasthurdle}
\alias{print.summary.fasthurdle}
\title{Print Method for Summary of Hurdle Models}
\usage{
\method{print}{summary.fasthurdle}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class "summary.fasthurdle".}

\item{digits}{Number of significant digits to use for printing.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
The summary object (invisibly).
}
\description{
Print a summary of a fitted hurdle model.
}
