% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{process_offsets}
\alias{process_offsets}
\title{Process Offsets for Hurdle Models}
\usage{
process_offsets(mf, mtX, mtZ, n)
}
\arguments{
\item{mf}{Model frame}

\item{mtX}{Count model terms}

\item{mtZ}{Zero model terms}

\item{n}{Number of observations}
}
\value{
List with offsetx and offsetz
}
\description{
Process Offsets for Hurdle Models
}
\keyword{internal}
