% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{process_weights}
\alias{process_weights}
\title{Process Weights for Hurdle Models}
\usage{
process_weights(mf, n)
}
\arguments{
\item{mf}{Model frame}

\item{n}{Number of observations}
}
\value{
Processed weights vector
}
\description{
Process Weights for Hurdle Models
}
\keyword{internal}
