% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{residuals.fasthurdle}
\alias{residuals.fasthurdle}
\title{Extract Residuals from a Hurdle Model}
\usage{
\method{residuals}{fasthurdle}(object, type = c("pearson", "response"), ...)
}
\arguments{
\item{object}{A fitted model object of class "fasthurdle".}

\item{type}{Character string specifying the type of residuals. "pearson" returns
Pearson residuals, and "response" returns response residuals.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A vector of residuals.
}
\description{
Extract residuals from a fitted hurdle model.
}
