% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{summary.fasthurdle}
\alias{summary.fasthurdle}
\title{Summary Method for Hurdle Models}
\usage{
\method{summary}{fasthurdle}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class "fasthurdle".}

\item{...}{Additional arguments (currently ignored).}
}
\value{
An object of class "summary.fasthurdle".
}
\description{
Compute a summary of a fitted hurdle model.
}
