% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{terms.fasthurdle}
\alias{terms.fasthurdle}
\title{Extract Terms from a Hurdle Model}
\usage{
\method{terms}{fasthurdle}(x, model = c("count", "zero"), ...)
}
\arguments{
\item{x}{A fitted model object of class "fasthurdle".}

\item{model}{Character string specifying which model terms to extract.
"count" extracts count model terms, and "zero" extracts zero hurdle model terms.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A terms object.
}
\description{
Extract the terms object from a fitted hurdle model.
}
