% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasthurdle.R
\name{validate_data}
\alias{validate_data}
\title{Validate Input Data for Hurdle Models}
\usage{
validate_data(Y, Z, zero.dist)
}
\arguments{
\item{Y}{Response variable}

\item{Z}{Zero model matrix}

\item{zero.dist}{Zero distribution type}
}
\value{
NULL, but stops with error if validation fails
}
\description{
Validate Input Data for Hurdle Models
}
\keyword{internal}
